malevnc: Support for Working with Janelia FlyEM Male VNC Dataset

Provides support for analysis of the male Drosophila ventral nerve cord dataset released in June 2023 by the Janelia FlyEM team in collaboration with the Hess lab, Google and Drosophila Connectomics Group, Cambridge. This includes reading of connectivity information, skeleton and mesh representations of neurons and symmetrising and inter-template registrations.

Registration

For information about left-right mirroring and symmetrising registrations within MANC, please see mirror_manc.

For information about across-template registrations, please see download_manc_registrations.

Package Options

malevnc.dataset A shortname defining the active dataset (usuallyMANC). See choose_malevnc_dataset for details.

malevnc.server The https URL to the main Janelia server hosting male VNC data.

See also

Author

Maintainer: Gregory Jefferis jefferis@gmail.com (ORCID)

Authors:

Dominik Krzeminski dkk33@cam.ac.uk (ORCID)

Examples

# List all package options
if (FALSE) { # \dontrun{
options()[grepl("^malevnc", names(options()))]
} # }