Read point annotations from DVID using neuprint authentication

Read point annotations from DVID using neuprint authentication

manc_user_annotations(
  email = "jefferis@gmail.com",
  node = "clio",
  raw = FALSE,
  simplifyVector = TRUE,
  neuprint_connection = NULL,
  tz = "UTC"
)

Arguments

email

The google email address used to authenticate with neuprint

node

The DVID node identifier

raw

Whether to return the raw httr::GET response (default FALSE) so that you can process it yourself rather than a pre-processed R list.

simplifyVector

Whether to simplify lists to vectors (and data frames where appropriate). Default TRUE, see jsonlite::fromJSON for details.

neuprint_connection

A neuprintr connection object returned by neuprint_login. This includes the required authorisation information to connect to DVID.

tz

Time zone for edit timestamps. Defaults to "UTC" i.e. Universal Time, Coordinated. Set to "" for your current timezone. See as.POSIXct for more details.

Value

A data.frame

Examples

# \donttest{
df=manc_user_annotations()
head(df)
#> $timestamp
#> POSIXct of length 0
#> 
json=httr::content(manc_user_annotations(raw=TRUE))
# }