Return clio-store body annotations for set of ids or a flexible query

clio_fields returns all the fields currently known to exist in the clio store for a given dataset. A short description...

Usage

manc_body_annotations(
  ids = NULL,
  query = NULL,
  json = FALSE,
  config = NULL,
  cache = FALSE,
  update.bodyids = FALSE,
  test = FALSE,
  show.extra = c("none", "user", "time", "all"),
  ...
)

clio_fields(dataset = getOption("malevnc.dataset"))

Arguments

ids

A set of body ids in any form understandable to manc_ids

query

A json query string (see examples or documentation) or an R list with field names as elements.

json

Whether to return unparsed JSON rather than an R list (default FALSE).

config

An optional httr::config (expert use only, must include a bearer token)

cache

Whether to cache the result of this call for 5 minutes.

update.bodyids

Whether to update the bodyid associated with annotations based on the position field. The default value of this has been switched to FALSE as of Feb 2022.

test

Whether to unset the clio-store test server (default FALSE)

show.extra

Extra columns to show with user/timestamp information.

...

Additional arguments passed to pbapply::pblapply

dataset

short name of the dataset (VNC or CNS)

Value

An R data.frame or a character vector containing JSON (when json=TRUE). Two additional fields will be added

original.bodyid When update.bodyids=TRUE this field contains the original bodyid from Clio whereas bodyid contains the updated value implied by the position.

auto TRUE signals that the record contains only data automatically copied over from DVID without any manual annotation.

See slack for details of the position / position type fields.

Details

Missing values in each output column are filled with NA. But if a whole column is missing from the results of a particular query then it will not appear at all.

When neither query and ids are missing then we return all entries in the clio store database. This currently includes annotations for all body ids - even the ones that are no longer current.

See also

Examples

if (FALSE) { # \dontrun{
manc_body_annotations(ids=11442)
manc_body_annotations(ids=11442, show.extra='user')
manc_body_annotations(query='{"hemilineage": "0B"}')
manc_body_annotations(query=list(user="janedoe@gmail.com"))
manc_body_annotations(query=list(soma_side="RHS"))
manc_body_annotations(ids=manc_xyz2bodyid(mancneckseeds))
# use clio node to ensure for bodyid consistency
manc_body_annotations(ids=
  manc_xyz2bodyid(mancneckseeds, node="clio"))

# fetch all bodyids
mba=manc_body_annotations()
} # }
clio_fields()
#> Error in httr::parse_url(url): length(url) == 1 is not TRUE