Return clio-store body annotations for set of ids or a flexible query

Return clio-store body annotations for set of ids or a flexible query

manc_body_annotations(
  ids = NULL,
  query = NULL,
  json = FALSE,
  config = NULL,
  ...
)

Arguments

ids
A vector of one or more ids

query
A json query string (see examples or documentation) or an R list with field names as elements.

json
Whether to return unparsed JSON rather than an R list (default FALSE).

config
An optional httr::config (expert use only, must include a bearer token)

...
Additional arguments passed to pbapply::pblapply

Value

An R data.frame or a character vector containing JSON (when json=TRUE).

Details

Missing values in each output column are filled with NA. But if a whole column is missing from the results of a particular query then it will not appear at all.

See also

Examples

if (FALSE) {
manc_body_annotations(ids=11442)
manc_body_annotations(query='{"hemilineage": "0B"}')
manc_body_annotations(query=list(user="janedoe@gmail.com"))
manc_body_annotations(query=list(soma_side="RHS"))
manc_body_annotations(ids=manc_xyz2id(mancneckseeds))
}