Return all DVID body annotations

Return all DVID body annotations

manc_dvid_annotations(
  ids = NULL,
  node = "neutu",
  rval = c("data.frame", "list"),
  columns_show = NULL,
  cache = FALSE
)

Arguments

ids

A set of body ids in any form understandable to manc_ids

node

A DVID node as returned by manc_dvid_node. The default is to return the current active (unlocked) node being used through neutu.

rval

Whether to return a fully parsed data.frame (the default) or an R list. The data.frame is easier to work with but typically includes NAs for many values that would be missing in the list.

columns_show

Whether to show all columns, or just with '_user', or '_time' suffix. Accepted options are: 'user', 'time', 'all'.

cache

Whether to cache the result of this call for 5 minutes.

Value

A tibble containing with columns including

bodyid as a numeric value

status

user

naming_user

instance

status_user

comment

NB only one bodyid is used regardless of whether the key-value returned has 0, 1 or 2 bodyid fields. When the ids are specified, missing ids will have a row containing the bodyid in question and then all other columns will be NA.

Details

See this Slack post from Stuart Berg for details.

Note that the original api call was <rootuuid>:master, but I have now just changed this to <neutu-uuid> as returned by manc_dvid_node. This was because the range query stopped working 16 May 2021, probably because of a bad node.

Examples

# \donttest{
mdf=manc_dvid_annotations()
head(mdf)
#> # A tibble: 6 × 51
#>   bodyid class     descr…¹ entry…² exit_…³ group hemil…⁴ insta…⁵ last_…⁶ nt_ac…⁷
#>    <dbl> <chr>     <chr>   <chr>   <chr>   <int> <chr>   <chr>   <chr>     <dbl>
#> 1  10000 Descendi… Giant … CvC     "None"  10000 None    TBD     lisa.m… 0.514  
#> 2  10001 Local Se… NA      ProLN_R  NA        NA NA      TBD     lisa.m… 0.258  
#> 3  10002 Descendi… Giant … CvC     "None"  10000 None    TBD     lisa.m… 0.460  
#> 4  10003 Local Se… NA      ProLN_L  NA        NA NA      TBD     cheong… 0.642  
#> 5  10004 Intrinsi… 13B in… None    ""      10004 13B     TBD     billyj… 0.00390
#> 6  10007 Local Se… NA      AbN3_R   NA        NA NA      TBD     imaan.… 0.886  
#> # … with 41 more variables: nt_gaba_prob <dbl>, nt_glutamate_prob <dbl>,
#> #   nt_unknown_prob <dbl>, origin <chr>, position <chr>, position_type <chr>,
#> #   predicted_nt <chr>, predicted_nt_prob <dbl>, prefix <chr>,
#> #   root_position <chr>, root_side <chr>, status <chr>, subclass <chr>,
#> #   synonyms <chr>, systematic_type <chr>, target <chr>, type <chr>,
#> #   user <chr>, tag <chr>, soma_neuromere <chr>, soma_side <chr>,
#> #   reviewer <chr>, avg_location <chr>, birthtime <chr>, confidence <chr>, …
table(mdf$status)
#> 
#>                                   0.5assign                Anchor 
#>                   519                  1104                   212 
#>                Orphan            PRT Orphan Prelim Roughly traced 
#>                   152                   432                  5399 
#>        Primary Anchor             RT Orphan        Roughly traced 
#>                     5                   160                 17145 
#>        Sensory Anchor           Soma Anchor                Traced 
#>                   339                     3                     2 
#>           Unimportant 
#>                  1623 

manc_dvid_annotations('DNp01')
#> # A tibble: 2 × 51
#>   bodyid class     descr…¹ entry…² exit_…³ group hemil…⁴ insta…⁵ last_…⁶ nt_ac…⁷
#>    <dbl> <chr>     <chr>   <chr>   <chr>   <int> <chr>   <chr>   <chr>     <dbl>
#> 1  10000 Descendi… Giant … CvC     None    10000 None    TBD     lisa.m…   0.514
#> 2  10002 Descendi… Giant … CvC     None    10000 None    TBD     lisa.m…   0.460
#> # … with 41 more variables: nt_gaba_prob <dbl>, nt_glutamate_prob <dbl>,
#> #   nt_unknown_prob <dbl>, origin <chr>, position <chr>, position_type <chr>,
#> #   predicted_nt <chr>, predicted_nt_prob <dbl>, prefix <chr>,
#> #   root_position <chr>, root_side <chr>, status <chr>, subclass <chr>,
#> #   synonyms <chr>, systematic_type <chr>, target <chr>, type <chr>,
#> #   user <chr>, tag <chr>, soma_neuromere <chr>, soma_side <chr>,
#> #   reviewer <chr>, avg_location <chr>, birthtime <chr>, confidence <chr>, …

if (FALSE) {
# compare live body annotations with version in clio
mdf.clio=manc_dvid_annotations('clio')
waldo::compare(mdf.clio, mdf)
}
# }