Return all DVID body annotations

Return all DVID body annotations

manc_dvid_annotations(
  node = "neutu",
  rval = c("data.frame", "list"),
  cache = FALSE
)

Arguments

node
A DVID node as returned by manc_dvid_node. The default is to return the current active (unlocked) node being used through neutu.

rval
Whether to return a fully parsed data.frame (the default) or an R list. The data.frame is easier to work with but typically includes NAs for many values that would be missing in the list.

cache
Whether to cache the result of this call for 5 minutes.

Value

A tibble containing with columns including

bodyid

status

user

naming_user

instance

status_user

comment

NB these are slightly modified from

Details

See this Slack post from Stuart Berg for details.

Note that the original api call was <rootuuid>:master, but I have now just changed this to <neutu-uuid> as returned by manc_dvid_node. This was because the range query stopped working 16 May 2021, probably because of a bad node.

Examples

# \donttest{
mdf=manc_dvid_annotations()
head(mdf)

#> # A tibble: 6 x 8
#>   bodyid class    status         user   naming_user instance status_user comment
#>    <dbl> <chr>    <chr>          <chr>  <chr>       <chr>    <chr>       <chr>  
#> 1  10000 Giant F… Prelim Roughl… takem… takemuras   10000_R  NA          NA     
#> 2 100000 NA       Prelim Roughl… lohffa takemuras   100000_R lohffa      NA     
#> 3 100001 NA       Anchor         NA     NA          NA       NA          NA     
#> 4 100002 NA       Prelim Roughl… lohffa NA          NA       lohffa      NA     
#> 5 100003 NA       NA             okeom… NA          NA       okeoman     NA     
#> 6 100009 NA       0.5assign      NA     NA          NA       NA          NA     

table(mdf$status)

#> 
#>             0.5assign                Anchor                Leaves 
#>                  4119                  1015                     1 
#>                Orphan Prelim Roughly traced        Primary Anchor 
#>                   311                 10093                    36 
#>            PRT Orphan        Sensory Anchor           Soma Anchor 
#>                     4                  2105                  9121 
#>           Unimportant 
#>                   707 

if (FALSE) {
# compare live body annotations with version in clio
mdf.clio=manc_dvid_annotations('clio')
waldo::compare(mdf.clio, mdf)
}
# }