Return full metadata from Clio/DVID for MANC bodyids (cached by default)

Return full metadata from Clio/DVID for MANC bodyids (cached by default)

manc_meta(
  ids = NULL,
  cache = TRUE,
  unique = FALSE,
  node = "neutu",
  update.bodyids = FALSE
)

Arguments

ids

bodyids specified in any form accepted by manc_ids

cache

Whether to use a cache with a 5 min lifetime, default TRUE.

unique

Whether to remove any duplicate ids from the results.

node

A DVID node as returned by manc_dvid_node. The default is to return the current active (unlocked) node being used through neutu.

update.bodyids

Whether to update the bodyid associated with annotations based on the position field. The default value of this has been switched to FALSE as of Feb 2022.

Value

A data.frame with columns determined by

manc_dvid_annotations and

manc_body_annotations. All columns from

manc_dvid_annotations will have the prefix dvid_ (except for the single bodyid column). In addition a column dvid_group

will be generated by parsing the dvid_instance column. This will have more up to date LR match information than the clio group field.

Details

This function will the latest per bodyid metadata using manc_dvid_annotations and manc_body_annotations. Since obtaining this information can take 5-10s even for a single id, it is by default cached with a 5 min timeout. The second call for the same request should be more or less instantaneous, while a different set of ids should still return much faster. Annotations are returned for the 'neutu' i.e. latest DVID node by default.

See also

Examples

if (FALSE) {
manc_meta('Giant Fiber')
manc_meta(10002)
}