Fetch neuprint metadata for MANC neurons

Fetch neuprint metadata for MANC neurons

manc_neuprint_meta(
  ids = NULL,
  conn = manc_neuprint(),
  roiInfo = FALSE,
  fields.regex.exclude = NULL,
  fields.regex.include = NULL,
  ...
)

Arguments

ids

A set of body ids in any form understandable to manc_ids

conn

Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

roiInfo

whether to include the roiInfo field detailing synapse numbers in different locations. This is omitted by default as it is returned as a character vector of unprocessed JSON.

fields.regex.exclude, fields.regex.include

Optional regular expressions to define fields to include or exclude from the returned metadata.

...

Additional arguments passed to neuprint_get_meta

Value

A data.frame with one row for each (unique) id and NAs for all columns except bodyid when neuprint holds no metadata.

See also

Examples

# \donttest{
manc_neuprint_meta("DNp01")
#>   bodyid  pre post upstream downstream synweight      voxels ntGabaProb
#> 1  10000 1033 2002     2002       4318      6320 38743961712  0.1707151
#> 2  10002  934 1902     1902       3826      5728 39414880927  0.1434814
#>   ntAcetylcholineProb ntGlutamateProb ntUnknownProb  type status    statusLabel
#> 1           0.5142120       0.2175471    0.09752578 DNp01 Traced Roughly traced
#> 2           0.4596285       0.2562304    0.14065966 DNp01 Traced Roughly traced
#>   cropped systematicType hemilineage somaSide description             class
#> 1   FALSE        DNlt002        None     <NA> Giant fiber descending neuron
#> 2   FALSE        DNlt002        None     <NA> Giant fiber descending neuron
#>   subclass group serial entryNerve          position somaNeuromere birthtime
#> 1       lt 10000     NA        CvC 24481,36044,67070          <NA>      <NA>
#> 2       lt 10000     NA        CvC 23217,35252,67070          None      <NA>
#>   somaLocation longTract predictedNtProb   predictedNt origin  target prefix
#> 1         <NA>      none       0.5142120 acetylcholine  brain LTct_RL     DN
#> 2         <NA>      none       0.4596285 acetylcholine  brain LTct_RL     DN
#>   positionType avgLocation serialMotif name exitNerve  tag rootSide modality
#> 1         neck        <NA>        <NA>  TBD      None <NA>      RHS     <NA>
#> 2         neck        <NA>        <NA>  TBD      None <NA>      LHS     <NA>
#>   subcluster transmission      rootPosition        synonyms tosomaPosition
#> 1         NA   electrical 24481,36044,67070 GF, Giant Fiber           <NA>
#> 2         NA   electrical 23217,35252,67070 GF, Giant Fiber           <NA>
#>   somaRadius source receptorType  soma
#> 1         NA   <NA>         <NA> FALSE
#> 2         NA   <NA>         <NA> FALSE

# use of a full CYPHER query. NB Each field relating to the neuron must be
# be preceded by "n."
bignogroup <-
  manc_neuprint_meta("where:NOT exists(n.group) AND n.synweight>5000 AND n.class CONTAINS 'neuron'")
head(bignogroup)
#>   bodyid  pre post upstream downstream synweight     voxels  ntGabaProb
#> 1  14487  523 4066     4066       3657      7723 1171534908 0.009269903
#> 2  17685 1618 2663     2663       8431     11094 2980016603 0.470944208
#> 3  14318  358 2589     2589       2534      5123  776162855 0.918771574
#> 4  20000  619 2330     2330       4410      6740  945138540 0.842403445
#> 5  11537  979 3288     3288       8739     12027 2371866344 0.792865791
#> 6  42768  643  445      445       6369      6814  370524828 0.091342640
#>   ntAcetylcholineProb ntGlutamateProb ntUnknownProb     type status
#> 1         0.957422450      0.02955370   0.003753942     <NA> Traced
#> 2         0.169709469      0.31638422   0.042962105     <NA> Traced
#> 3         0.003175267      0.07543413   0.002619032 AN05B013 Traced
#> 4         0.019575450      0.12935726   0.008663844 AN05B022 Traced
#> 5         0.017601760      0.18710177   0.002430678 AN09A006 Traced
#> 6         0.153858850      0.74405113   0.010747370     <NA> Traced
#>      statusLabel cropped systematicType hemilineage somaSide
#> 1 Roughly traced   FALSE           <NA>         TBD      RHS
#> 2 Roughly traced   FALSE           <NA>         TBD  Midline
#> 3 Roughly traced   FALSE       AN05B013         05B      LHS
#> 4 Roughly traced   FALSE       AN05B022         05B      RHS
#> 5 Roughly traced   FALSE       AN09A006         09A      RHS
#> 6 Roughly traced   FALSE           <NA>         TBD      LHS
#>                                                                                             description
#> 1                                                                                                  <NA>
#> 2                                                                    dorsal soma; possibly serial 10477
#> 3                                                                            5B in T1 LHS; 1st pass ECM
#> 4 5B in T1 RHS; needs major merge from 19809, 37303, 44681; truncated by tissue damage at posterior end
#> 5                                                                                 9A? primary in T2 RHS
#> 6                                                                                                  <NA>
#>              class subclass group serial entryNerve          position
#> 1 ascending neuron       BA    NA     NA       None 23293,38376,67070
#> 2 ascending neuron       BA    NA  13750       <NA> 24265,31026,57936
#> 3 ascending neuron       BA    NA     NA       None 18930,37472,45357
#> 4 ascending neuron       BA    NA     NA       None 25195,38150,45000
#> 5 ascending neuron       BA    NA  10628       None 36674,31985,43673
#> 6 intrinsic neuron       BI    NA     NA       <NA> 25729,13391,20666
#>   somaNeuromere birthtime      somaLocation longTract predictedNtProb
#> 1            A8   primary  29113,10153,5684      <NA>       0.9574225
#> 2            T1   primary 24265,31026,57936      <NA>       0.4709442
#> 3            T1 secondary 19687,39944,44160                 0.9187716
#> 4            T1 secondary 23325,41413,44025                 0.8424034
#> 5            T2   primary 36674,31985,43673      <NA>       0.7928658
#> 6            A7   primary  20541,17948,3759      <NA>       0.7440511
#>     predictedNt        origin        target prefix positionType
#> 1 acetylcholine         multi         multi     BA         user
#> 2          gaba    ANm.IntTct    ANm.IntTct     BA         soma
#> 3          gaba         multi         multi     BA         user
#> 4          gaba         multi         multi     BA         user
#> 5          gaba         multi       LegNp_R     BA         soma
#> 6     glutamate ANm.LegNpT3_L ANm.LegNpT3_L     BI         auto
#>                                                                                                                                                                                 avgLocation
#> 1                                                                                                                                                                                        NA
#> 2                                                                                                                                                                                        NA
#> 3 list(18873, 37472, 45448),list(name = "cartesian-3d", properties = list(href = "http://spatialreference.org/ref/sr-org/9157/ogcwkt/", type = "ogcwkt"), srid = 9157, type = "link"),Point
#> 4 list(25171, 38150, 45076),list(name = "cartesian-3d", properties = list(href = "http://spatialreference.org/ref/sr-org/9157/ogcwkt/", type = "ogcwkt"), srid = 9157, type = "link"),Point
#> 5                                                                                                                                                                                        NA
#> 6                                                                                                                                                                                        NA
#>   serialMotif          name exitNerve  tag rootSide modality subcluster
#> 1        <NA>           TBD       CvC <NA>      RHS     <NA>         NA
#> 2      dorsal           TBD       CvC <NA>  Midline     <NA>         NA
#> 3        <NA> AN05B013_T1_L       CvC <NA>      RHS     <NA>      13169
#> 4        <NA> AN05B022_T1_R       CvC <NA>      LHS     <NA>         NA
#> 5     complex AN09A006_T2_R       CvC <NA>      RHS     <NA>         NA
#> 6        <NA>           TBD      <NA> <NA>     <NA>     <NA>         NA
#>              transmission rootPosition synonyms tosomaPosition somaRadius
#> 1                    <NA>         <NA>     <NA>           <NA>         NA
#> 2 putative neurosecretory         <NA>     <NA>           <NA>         NA
#> 3                    <NA>         <NA>     <NA>           <NA>         NA
#> 4                    <NA>         <NA>     <NA>           <NA>         NA
#> 5                    <NA>         <NA>     <NA>           <NA>         NA
#> 6                    <NA>         <NA>     <NA>           <NA>         NA
#>   source receptorType soma
#> 1   <NA>         <NA> TRUE
#> 2   <NA>         <NA> TRUE
#> 3   <NA>         <NA> TRUE
#> 4   <NA>         <NA> TRUE
#> 5   <NA>         <NA> TRUE
#> 6   <NA>         <NA> TRUE
# }