Fetch neuprint metadata for MANC neurons

Fetch neuprint metadata for MANC neurons

manc_neuprint_meta(ids = NULL, conn = manc_neuprint(), roiInfo = FALSE)

Arguments

ids

A set of body ids in any form understandable to manc_ids

conn

Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

roiInfo

whether to include the roiInfo field detailing synapse numbers in different locations. This is omitted by default as it is returned as a character vector of unprocessed JSON.

Value

A data.frame with one row for each (unique) id and NAs for all columns except bodyid when neuprint holds no metadata.

See also

Examples

# \donttest{
manc_neuprint_meta("DNp01")
#>   bodyid  pre post upstream downstream synweight status    statusLabel
#> 1  10000 1033 2016     2016       5150      7166 Traced Roughly traced
#> 2  10002  934 1948     1948       4606      6554 Traced Roughly traced
#>        voxels ntGabaProb ntAcetylcholineProb ntGlutamateProb ntUnknownProb name
#> 1 38743961712  0.1707151           0.5142120       0.2175471    0.09752578  TBD
#> 2 39414880927  0.1434814           0.4596285       0.2562304    0.14065966  TBD
#>            position predictedNtProb   predictedNt positionType cropped
#> 1 24481,36044,67070       0.5142120 acetylcholine         neck   FALSE
#> 2 23217,35252,67070       0.4596285 acetylcholine         neck   FALSE
#>        class   subclass group rootSide entryNerve origin  target prefix
#> 1 Descending electrical 10000      RHS        CvC  brain LTct_RL     DN
#> 2 Descending electrical 10000      LHS        CvC  brain LTct_RL     DN
#>   reviewer  type systematicType hemilineage somaSide serial exitNerve
#> 1     <NA> DNp01        DNlt003        None     <NA>     NA      None
#> 2     <NA> DNp01        DNlt003        None     <NA>     NA      None
#>   somaNeuromere birthtime somaLocation subcluster                 user
#> 1          <NA>      <NA>         <NA>         NA Gregory SXE Jefferis
#> 2          None      <NA>         <NA>         NA    Elizabeth C Marin
#>      lastModifiedBy description groupOld  tag      rootPosition oldBodyids
#> 1 Elizabeth C Marin Giant fiber       NA <NA> 24481,36044,67070         NA
#> 2 Elizabeth C Marin Giant fiber       NA <NA> 23217,35252,67070         NA
#>   avgLocation confidence typingNotes    synonyms somaRadius tosomaPosition
#> 1        <NA>       <NA>        <NA> Giant Fiber         NA           <NA>
#> 2        <NA>       <NA>        <NA> Giant Fiber         NA           <NA>
#>   source modality receptorType  soma
#> 1   <NA>     <NA>         <NA> FALSE
#> 2   <NA>     <NA>         <NA> FALSE

# use of a full CYPHER query. NB Each field relating to the neuron must be
# be preceded by "n."
bignogroup <-
  manc_neuprint_meta("where:NOT exists(n.group) AND n.synweight>5000 AND n.class CONTAINS 'neuron'")
head(bignogroup)
#>   bodyid pre post upstream downstream synweight status           statusLabel
#> 1  14487 523 4599     4599       4301      8900 Traced        Roughly traced
#> 2  20000 619 2721     2721       5081      7802 Traced        Roughly traced
#> 3  11537 979 3769     3769       9673     13442 Traced Prelim Roughly traced
#> 4  17733 340 2749     2749       2752      5501 Traced        Roughly traced
#> 5  16084 782 2353     2353       6987      9340 Traced        Roughly traced
#> 6  15330 703 2619     2619       6520      9139 Traced Prelim Roughly traced
#>       voxels  ntGabaProb ntAcetylcholineProb ntGlutamateProb ntUnknownProb name
#> 1 1171534908 0.009269903         0.957422450      0.02955370   0.003753942  TBD
#> 2  945138540 0.842403445         0.019575450      0.12935726   0.008663844  TBD
#> 3 2371866344 0.792865791         0.017601760      0.18710177   0.002430678  TBD
#> 4  743873684 0.883237927         0.005398709      0.10719202   0.004171349  TBD
#> 5  788893095 0.872685350         0.020873548      0.10333776   0.003103343  TBD
#> 6 1006877150 0.008896410         0.941923661      0.04556721   0.003612722  TBD
#>            position predictedNtProb   predictedNt positionType cropped
#> 1 23293,38376,67070       0.9574225 acetylcholine         user   FALSE
#> 2 25195,38150,45000       0.8424034          gaba         user   FALSE
#> 3 36674,31985,43673       0.7928658          gaba         soma   FALSE
#> 4 18951,37472,45337       0.8832379          gaba         user   FALSE
#> 5  21482,17876,2510       0.8726854          gaba         soma   FALSE
#> 6  27535,16485,3922       0.9419237 acetylcholine         soma   FALSE
#>                   class subclass group rootSide entryNerve origin  target
#> 1 Ascending Interneuron     <NA>    NA      RHS       None  multi   multi
#> 2 Ascending Interneuron     <NA>    NA      LHS       None  multi   multi
#> 3 Ascending Interneuron     <NA>    NA      RHS       None  multi LegNp_R
#> 4 Ascending Interneuron     <NA>    NA      RHS       None  multi   multi
#> 5      Intrinsic neuron     <NA>    NA     <NA>       <NA>  ANm_L   ANm_R
#> 6      Intrinsic neuron     <NA>    NA     <NA>       <NA>  ANm_R   ANm_L
#>   prefix reviewer type systematicType hemilineage somaSide serial exitNerve
#> 1     BA     <NA> <NA>           <NA>         TBD      RHS     NA       CvC
#> 2     BA     <NA> <NA>           <NA>         05B      RHS     NA       CvC
#> 3     BA     <NA> <NA>           <NA>         09A      RHS  10628       CvC
#> 4     BA     <NA> <NA>           <NA>         05B      LHS     NA       CvC
#> 5     BR     <NA> <NA>           <NA>         TBD      LHS     NA      <NA>
#> 6     BR     <NA> <NA>           <NA>         TBD      RHS     NA      <NA>
#>   somaNeuromere      birthtime      somaLocation subcluster
#> 1            A8           <NA>  29113,10153,5684         NA
#> 2            T1 secondary auto 23325,41413,44025         NA
#> 3            T2        primary 36674,31985,43673         NA
#> 4            T1      secondary 20877,41021,44461         NA
#> 5            A9 secondary auto  21482,17876,2510         NA
#> 6            A8 secondary auto  27535,16485,3922      15330
#>                   user    lastModifiedBy
#> 1 Gregory SXE Jefferis    David Shepherd
#> 2    Elizabeth C Marin      Imaan Tamimi
#> 3    Elizabeth C Marin Elizabeth C Marin
#> 4    Elizabeth C Marin    Marina Gkantia
#> 5    Elizabeth C Marin Elizabeth C Marin
#> 6          Marta Costa Elizabeth C Marin
#>                                                                                           description
#> 1                                                                                                <NA>
#> 2 5B in T1 RHS; needs major merge from 19809 37303 44681; truncated by tissue damage at posterior end
#> 3                                                                               9A? primary in T2 RHS
#> 4                                                                          5B in T1 LHS; 1st pass ECM
#> 5                                                                                    was: group 16084
#> 6                                                                                                <NA>
#>   groupOld  tag rootPosition oldBodyids
#> 1       NA <NA>         <NA>         NA
#> 2    14318 <NA>         <NA>         NA
#> 3       NA <NA>         <NA>         NA
#> 4    14318 <NA>         <NA>         NA
#> 5       NA <NA>         <NA>         NA
#> 6       NA <NA>         <NA>         NA
#>                                                                                                                                                                                 avgLocation
#> 1                                                                                                                                                                                        NA
#> 2 list(25171, 38150, 45076),list(name = "cartesian-3d", properties = list(href = "http://spatialreference.org/ref/sr-org/9157/ogcwkt/", type = "ogcwkt"), srid = 9157, type = "link"),Point
#> 3                                                                                                                                                                                        NA
#> 4 list(18873, 37472, 45448),list(name = "cartesian-3d", properties = list(href = "http://spatialreference.org/ref/sr-org/9157/ogcwkt/", type = "ogcwkt"), srid = 9157, type = "link"),Point
#> 5                                                                                                                                                                                        NA
#> 6                                                                                                                                                                                        NA
#>   confidence                              typingNotes synonyms somaRadius
#> 1       <NA>                                     <NA>     <NA>         NA
#> 2          3 ANs; Does not include putative primaries     <NA>         NA
#> 3       <NA>                                     <NA>     <NA>         NA
#> 4          3 ANs; Does not include putative primaries     <NA>         NA
#> 5       <NA>                                     <NA>     <NA>         NA
#> 6       <NA>                                     <NA>     <NA>         NA
#>   tosomaPosition source modality receptorType soma
#> 1           <NA>   <NA>     <NA>         <NA> TRUE
#> 2           <NA>   <NA>     <NA>         <NA> TRUE
#> 3           <NA>   <NA>     <NA>         <NA> TRUE
#> 4           <NA>   <NA>     <NA>         <NA> TRUE
#> 5           <NA>   <NA>     <NA>         <NA> TRUE
#> 6           <NA>   <NA>     <NA>         <NA> TRUE
# }