Fetch neuprint metadata for MANC neurons

Fetch neuprint metadata for MANC neurons

Usage

manc_neuprint_meta(
  ids = NULL,
  conn = manc_neuprint(),
  roiInfo = FALSE,
  fields.regex.exclude = NULL,
  fields.regex.include = NULL,
  ...
)

Arguments

ids

A set of body ids in any form understandable to manc_ids

conn

Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

roiInfo

whether to include the roiInfo field detailing synapse numbers in different locations. This is omitted by default as it is returned as a character vector of unprocessed JSON.

fields.regex.exclude, fields.regex.include

Optional regular expressions to define fields to include or exclude from the returned metadata.

...

Additional arguments passed to neuprint_get_meta

Value

A data.frame with one row for each (unique) id and NAs for all columns except bodyid when neuprint holds no metadata.

See also

Examples

# \donttest{
manc_neuprint_meta("DNp01")
#>   bodyid post  pre downstream upstream synweight ntGabaProb ntAcetylcholineProb
#> 1  10000 2059 1033       4318     2059      6377  0.1707151           0.5142119
#> 2  10002 1973  934       3826     1973      5799  0.1434814           0.4596285
#>   ntGlutamateProb ntUnknownProb   predictedNt      voxels             class
#> 1       0.2175471    0.09752578 acetylcholine 38743961712 descending neuron
#> 2       0.2562304    0.14065966 acetylcholine 39414880927 descending neuron
#>   entryNerve name origin subcluster          location locationType
#> 1        CvC  TBD  brain         NA 24481,36044,67070         neck
#> 2        CvC  TBD  brain         NA 23217,35252,67070         neck
#>   predictedNtProb prefix rootSide    statusLabel subclass        synonyms
#> 1       0.5142119     DN      RHS Roughly traced       lt GF, Giant Fiber
#> 2       0.4596285     DN      LHS Roughly traced       lt GF, Giant Fiber
#>   systematicType  target  type        vfbId modality  tag cluster receptorType
#> 1        DNlt002 LTct_RL DNp01 VFB_jrcv07ps     <NA> <NA>      NA         <NA>
#> 2        DNlt002 LTct_RL DNp01 VFB_jrcv07pu     <NA> <NA>      NA         <NA>
#>   status group somaSide birthtime hemilineage serial serialMotif somaNeuromere
#> 1 Traced 10000     <NA>      <NA>        <NA>     NA        <NA>          <NA>
#> 2 Traced 10000     <NA>      <NA>        <NA>     NA        <NA>          <NA>
#>   somaLocation description avgLocation exitNerve subclassabbr transmission
#> 1         <NA> Giant fiber        <NA>      None         <NA>   electrical
#> 2         <NA> Giant fiber        <NA>      None         <NA>   electrical
#>   tosomaLocation longTract      rootLocation source  soma
#> 1           <NA>      none 24481,36044,67070   <NA> FALSE
#> 2           <NA>      none 23217,35252,67070   <NA> FALSE

# use of a full CYPHER query. NB Each field relating to the neuron must be
# be preceded by "n."
bignogroup <-
  manc_neuprint_meta("where:NOT exists(n.group) AND n.synweight>5000 AND n.class CONTAINS 'neuron'")
head(bignogroup)
#>   bodyid post  pre downstream upstream synweight  ntGabaProb
#> 1  14487 4622  523       3657     4622      8279 0.009269903
#> 2  11537 3770  979       8739     3770     12509 0.792865800
#> 3  20000 2724  619       4410     2724      7134 0.537725900
#> 4  17733 2758  340       2410     2758      5168 0.883237900
#> 5  17685 3199 1618       8431     3199     11630 0.466351630
#> 6  14318 3022  358       2534     3022      5556 0.918771560
#>   ntAcetylcholineProb ntGlutamateProb ntUnknownProb   predictedNt     voxels
#> 1         0.957422440      0.02955370   0.003753942 acetylcholine 1171534908
#> 2         0.017601760      0.18710177   0.002430678          gaba 2371866344
#> 3         0.378509670      0.07122201   0.012542419          gaba  945138540
#> 4         0.005398709      0.10719202   0.004171349          gaba  743873684
#> 5         0.169925300      0.32034713   0.043375943          gaba 2980016603
#> 6         0.003175267      0.07543413   0.002619033          gaba  776162855
#>              class entryNerve          name     origin subcluster
#> 1 ascending neuron       None ANXXX196_A8_R      multi         NA
#> 2 ascending neuron       None AN09A007_T2_R      multi      10628
#> 3 ascending neuron       None AN05B047_T1_R      multi         NA
#> 4 ascending neuron       None AN05B058_T1_L      multi         NA
#> 5 ascending neuron       <NA> ANXXX034_T1_M ANm.IntTct         NA
#> 6 ascending neuron       None AN05B058_T1_L      multi      13169
#>            location locationType predictedNtProb prefix rootSide    statusLabel
#> 1 23293,38376,67070         user       0.9574224     BA      RHS Roughly traced
#> 2 36674,31985,43673         soma       0.7928658     BA      RHS Roughly traced
#> 3 25195,38150,45000         user       0.8424035     BA      LHS Roughly traced
#> 4 18951,37472,45337         user       0.8832379     BA      RHS Roughly traced
#> 5 24265,31026,57936         soma       0.4709442     BA  Midline Roughly traced
#> 6 18930,37472,45357         user       0.9187716     BA      RHS Roughly traced
#>   subclass synonyms systematicType     target     type        vfbId
#> 1       BA     <NA>       ANXXX196      multi ANXXX196 VFB_jrcv0b6f
#> 2       BA     <NA>       AN09A007    LegNp_R AN09A007 VFB_jrcv08wh
#> 3       BA     <NA>       AN05B047      multi AN05B047 VFB_jrcv0ffk
#> 4       BA     <NA>       AN05B058      multi AN05B058 VFB_jrcv0dol
#> 5       BA     <NA>       ANXXX034 ANm.IntTct ANXXX034 VFB_jrcv0dn9
#> 6       BA     <NA>       AN05B058      multi AN05B058 VFB_jrcv0b1q
#>                 modality  tag cluster receptorType status group somaSide
#> 1   chemosensory_unknown <NA>      NA         <NA> Traced    NA      RHS
#> 2                   <NA> <NA>      NA         <NA> Traced    NA      RHS
#> 3 proprioceptive_tactile <NA>      NA         <NA> Traced    NA      RHS
#> 4 proprioceptive_tactile <NA>      NA         <NA> Traced    NA      LHS
#> 5                   <NA> <NA>      NA         <NA> Traced    NA  Midline
#> 6 proprioceptive_tactile <NA>      NA         <NA> Traced    NA      LHS
#>   birthtime hemilineage serial serialMotif somaNeuromere      somaLocation
#> 1   primary         TBD     NA        <NA>            A8  29113,10153,5684
#> 2   primary         09A  10628     complex            T2 36674,31985,43673
#> 3 secondary         05B     NA        <NA>            T1 23325,41413,44025
#> 4 secondary         05B     NA        <NA>            T1 20877,41021,44461
#> 5   primary         TBD  13750      dorsal            T1 24265,31026,57936
#> 6 secondary         05B     NA        <NA>            T1 19687,39944,44160
#>                                                                                             description
#> 1                                                                                                  <NA>
#> 2                                                                                 9A? primary in T2 RHS
#> 3 5B in T1 RHS; needs major merge from 19809, 37303, 44681; truncated by tissue damage at posterior end
#> 4                                                                            5B in T1 LHS; 1st pass ECM
#> 5                                                                    dorsal soma; possibly serial 10477
#> 6                                                                            5B in T1 LHS; 1st pass ECM
#>                                                                                                                                                                                 avgLocation
#> 1                                                                                                                                                                                          
#> 2                                                                                                                                                                                          
#> 3 list(25171, 38150, 45076),list(name = "cartesian-3d", properties = list(href = "http://spatialreference.org/ref/sr-org/9157/ogcwkt/", type = "ogcwkt"), srid = 9157, type = "link"),Point
#> 4 list(18873, 37472, 45448),list(name = "cartesian-3d", properties = list(href = "http://spatialreference.org/ref/sr-org/9157/ogcwkt/", type = "ogcwkt"), srid = 9157, type = "link"),Point
#> 5                                                                                                                                                                                          
#> 6 list(18873, 37472, 45448),list(name = "cartesian-3d", properties = list(href = "http://spatialreference.org/ref/sr-org/9157/ogcwkt/", type = "ogcwkt"), srid = 9157, type = "link"),Point
#>   exitNerve subclassabbr            transmission tosomaLocation longTract
#> 1       CvC         <NA>                    <NA>           <NA>      <NA>
#> 2       CvC         <NA>                    <NA>           <NA>      <NA>
#> 3       CvC         <NA>                    <NA>           <NA>      <NA>
#> 4       CvC         <NA>                    <NA>           <NA>      <NA>
#> 5       CvC         <NA> putative neurosecretory           <NA>      <NA>
#> 6       CvC         <NA>                    <NA>           <NA>      <NA>
#>   rootLocation source soma
#> 1         <NA>   <NA> TRUE
#> 2         <NA>   <NA> TRUE
#> 3         <NA>   <NA> TRUE
#> 4         <NA>   <NA> TRUE
#> 5         <NA>   <NA> TRUE
#> 6         <NA>   <NA> TRUE
# }