Fetch neuprint metadata for MANC neurons

Fetch neuprint metadata for MANC neurons

Usage

manc_neuprint_meta(
  ids = NULL,
  conn = manc_neuprint(),
  roiInfo = FALSE,
  fields.regex.exclude = NULL,
  fields.regex.include = NULL,
  ...
)

Arguments

ids

A set of body ids in any form understandable to manc_ids

conn

Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

roiInfo

whether to include the roiInfo field detailing synapse numbers in different locations. This is omitted by default as it is returned as a character vector of unprocessed JSON.

fields.regex.exclude, fields.regex.include

Optional regular expressions to define fields to include or exclude from the returned metadata.

...

Additional arguments passed to neuprint_get_meta

Value

A data.frame with one row for each (unique) id and NAs for all columns except bodyid when neuprint holds no metadata.

See also

Examples

# \donttest{
manc_neuprint_meta("DNp01")
#>   bodyid  pre post upstream downstream synweight status    statusLabel cropped
#> 1  10000 1033 2002     2002       4318      6320 Traced Roughly traced   FALSE
#> 2  10000 1033 2002     2002       4318      6320 Traced Roughly traced   FALSE
#> 3  10000 1033 2002     2002       4318      6320 Traced Roughly traced   FALSE
#> 4  10002  934 1902     1902       3826      5728 Traced Roughly traced   FALSE
#> 5  10002  934 1902     1902       3826      5728 Traced Roughly traced   FALSE
#> 6  10002  934 1902     1902       3826      5728 Traced Roughly traced   FALSE
#>    type name somaSide hemilineage systematicType             class subclass
#> 1 DNp01  TBD     <NA>        <NA>        DNlt002 descending neuron       lt
#> 2 DNp01  TBD     <NA>        <NA>        DNlt002 descending neuron       lt
#> 3 DNp01  TBD     <NA>        <NA>        DNlt002 descending neuron       lt
#> 4 DNp01  TBD                 <NA>        DNlt002 descending neuron       lt
#> 5 DNp01  TBD                 <NA>        DNlt002 descending neuron       lt
#> 6 DNp01  TBD                 <NA>        DNlt002 descending neuron       lt
#>   group serial          position somaNeuromere birthtime somaLocation
#> 1 10000     NA 24481,36044,67070          <NA>      <NA>         <NA>
#> 2 10000     NA 24481,36044,67070          <NA>      <NA>         <NA>
#> 3 10000     NA 24481,36044,67070          <NA>      <NA>         <NA>
#> 4 10000     NA 23217,35252,67070          None      <NA>         <NA>
#> 5 10000     NA 23217,35252,67070          None      <NA>         <NA>
#> 6 10000     NA 23217,35252,67070          None      <NA>         <NA>
#>        voxels ntGabaProb ntAcetylcholineProb ntGlutamateProb ntUnknownProb
#> 1 38743961712  0.1707151           0.5142120       0.2175471    0.09752578
#> 2 38743961712  0.1707151           0.5142120       0.2175471    0.09752578
#> 3 38743961712  0.1707151           0.5142120       0.2175471    0.09752578
#> 4 39414880927  0.1434814           0.4596285       0.2562304    0.14065966
#> 5 39414880927  0.1434814           0.4596285       0.2562304    0.14065966
#> 6 39414880927  0.1434814           0.4596285       0.2562304    0.14065966
#>   predictedNtProb   predictedNt origin  target subcluster positionType
#> 1       0.5142120 acetylcholine  brain LTct_RL         NA         neck
#> 2       0.5142120 acetylcholine  brain LTct_RL         NA         neck
#> 3       0.5142120 acetylcholine  brain LTct_RL         NA         neck
#> 4       0.4596285 acetylcholine  brain LTct_RL         NA         neck
#> 5       0.4596285 acetylcholine  brain LTct_RL         NA         neck
#> 6       0.4596285 acetylcholine  brain LTct_RL         NA         neck
#>   serialMotif        synonyms rootSide entryNerve  tag modality longTract
#> 1        <NA> GF, Giant Fiber      RHS        CvC <NA>     <NA>      none
#> 2        <NA> GF, Giant Fiber      RHS        CvC <NA>     <NA>      none
#> 3        <NA> GF, Giant Fiber      RHS        CvC <NA>     <NA>      none
#> 4        <NA> GF, Giant Fiber      LHS        CvC <NA>     <NA>      none
#> 5        <NA> GF, Giant Fiber      LHS        CvC <NA>     <NA>      none
#> 6        <NA> GF, Giant Fiber      LHS        CvC <NA>     <NA>      none
#>   exitNerve transmission rootLocation tosomaLocation  soma
#> 1      None   electrical        24481             NA FALSE
#> 2      None   electrical        67070             NA FALSE
#> 3      None   electrical        35252             NA FALSE
#> 4      None   electrical        36044             NA FALSE
#> 5      None   electrical        23217             NA FALSE
#> 6      None   electrical        67070             NA FALSE

# use of a full CYPHER query. NB Each field relating to the neuron must be
# be preceded by "n."
bignogroup <-
  manc_neuprint_meta("where:NOT exists(n.group) AND n.synweight>5000 AND n.class CONTAINS 'neuron'")
head(bignogroup)
#>   bodyid  pre post upstream downstream synweight status    statusLabel cropped
#> 1  14487  523 4066     4066       3657      7723 Traced Roughly traced   FALSE
#> 2  17685 1618 2663     2663       8431     11094 Traced Roughly traced   FALSE
#> 3  14318  358 2589     2589       2534      5123 Traced Roughly traced   FALSE
#> 4  20000  619 2330     2330       4410      6740 Traced Roughly traced   FALSE
#> 5  11537  979 3288     3288       8739     12027 Traced Roughly traced   FALSE
#> 6  15330  703 2281     2281       5575      7856 Traced Roughly traced   FALSE
#>       type          name somaSide hemilineage systematicType            class
#> 1 ANXXX196 ANXXX196_A8_R      RHS         TBD       ANXXX196 ascending neuron
#> 2 ANXXX034 ANXXX034_T1_M  Midline         TBD       ANXXX034 ascending neuron
#> 3 AN05B058 AN05B058_T1_L      LHS         05B       AN05B058 ascending neuron
#> 4 AN05B047 AN05B047_T1_R      RHS         05B       AN05B047 ascending neuron
#> 5 AN09A008 AN09A008_T2_R      RHS         09A       AN09A008 ascending neuron
#> 6 INXXX234 INXXX234_A8_R      RHS         TBD       INXXX234 intrinsic neuron
#>   subclass group serial          position somaNeuromere birthtime
#> 1       BA    NA     NA 23293,38376,67070            A8   primary
#> 2       BA    NA  13750 24265,31026,57936            T1   primary
#> 3       BA    NA     NA 18930,37472,45357            T1 secondary
#> 4       BA    NA     NA 25195,38150,45000            T1 secondary
#> 5       BA    NA  10628 36674,31985,43673            T2   primary
#> 6       BR    NA     NA  27535,16485,3922            A8   primary
#>        somaLocation     voxels  ntGabaProb ntAcetylcholineProb ntGlutamateProb
#> 1  29113,10153,5684 1171534908 0.009269903         0.957422450      0.02955370
#> 2 24265,31026,57936 2980016603 0.470944208         0.169709469      0.31638422
#> 3 19687,39944,44160  776162855 0.918771574         0.003175267      0.07543413
#> 4 23325,41413,44025  945138540 0.842403445         0.019575450      0.12935726
#> 5 36674,31985,43673 2371866344 0.792865791         0.017601760      0.18710177
#> 6  27535,16485,3922 1006877150 0.008896410         0.941923661      0.04556721
#>   ntUnknownProb predictedNtProb   predictedNt     origin     target subcluster
#> 1   0.003753942       0.9574225 acetylcholine      multi      multi         NA
#> 2   0.042962105       0.4709442          gaba ANm.IntTct ANm.IntTct         NA
#> 3   0.002619032       0.9187716          gaba      multi      multi      13169
#> 4   0.008663844       0.8424034          gaba      multi      multi         NA
#> 5   0.002430678       0.7928658          gaba      multi    LegNp_R      10628
#> 6   0.003612722       0.9419237 acetylcholine      ANm_R      ANm_L      15330
#>   positionType serialMotif synonyms rootSide entryNerve  tag modality longTract
#> 1         user        <NA>     <NA>      RHS       None <NA>     <NA>      <NA>
#> 2         soma      dorsal     <NA>  Midline       <NA> <NA>     <NA>      <NA>
#> 3         user        <NA>     <NA>      RHS       None <NA>     <NA>          
#> 4         user        <NA>     <NA>      LHS       None <NA>     <NA>          
#> 5         soma     complex     <NA>      RHS       None <NA>     <NA>      <NA>
#> 6         soma        <NA>     <NA>     <NA>       <NA> <NA>     <NA>      <NA>
#>   exitNerve            transmission rootLocation tosomaLocation soma
#> 1       CvC                    <NA>           NA             NA TRUE
#> 2       CvC putative neurosecretory           NA             NA TRUE
#> 3       CvC                    <NA>           NA             NA TRUE
#> 4       CvC                    <NA>           NA             NA TRUE
#> 5       CvC                    <NA>           NA             NA TRUE
#> 6      <NA>                    <NA>           NA             NA TRUE
# }