Fetch neuprint metadata for MANC neurons

Fetch neuprint metadata for MANC neurons

Usage

manc_neuprint_meta(
  ids = NULL,
  conn = manc_neuprint(),
  roiInfo = FALSE,
  fields.regex.exclude = NULL,
  fields.regex.include = NULL,
  ...
)

Arguments

ids

A set of body ids in any form understandable to manc_ids

conn

Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

roiInfo

whether to include the roiInfo field detailing synapse numbers in different locations. This is omitted by default as it is returned as a character vector of unprocessed JSON.

fields.regex.exclude, fields.regex.include

Optional regular expressions to define fields to include or exclude from the returned metadata.

...

Additional arguments passed to neuprint_get_meta

Value

A data.frame with one row for each (unique) id and NAs for all columns except bodyid when neuprint holds no metadata.

Details

When ids = NULL then a default set of bodies is selected using the manc_dvid_annotations function. Since April 2025 this uses the node='neuprint'. This should correspond to all neurons with an annotation. You can also use other searches e.g. to fetch all neurons, see examples.

See also

Examples

# \donttest{
manc_neuprint_meta("DNp01")
#>   bodyid post  pre downstream upstream synweight totalNtPredictions
#> 1  10000 2002 1027       4318     2002      6320               1027
#> 2  10002 1902  933       3826     1902      5728                933
#>     predictedNt celltypeTotalNtPredictions celltypePredictedNt      voxels
#> 1 acetylcholine                       1960       acetylcholine 38743961712
#> 2 acetylcholine                       1960       acetylcholine 39414880927
#>               class entryNerve          name ntAcetylcholineProb ntGabaProb
#> 1 descending neuron        CvC DNlt002_CvC_R           0.5142120  0.1707151
#> 2 descending neuron        CvC DNlt002_CvC_L           0.4596285  0.1434814
#>   ntGlutamateProb ntUnknownProb origin          location locationType
#> 1       0.2175471    0.09752578  brain 24481,36044,67070         neck
#> 2       0.2562304    0.14065966  brain 23217,35252,67070         neck
#>   predictedNtProb prefix      rootLocation rootSide    statusLabel subclass
#> 1       0.5142120     DN 24481,36044,67070      RHS Roughly traced       lt
#> 2       0.4596285     DN 23217,35252,67070      LHS Roughly traced       lt
#>          synonyms systematicType  target  type        vfbId modality  tag
#> 1 GF, Giant Fiber        DNlt002 LTct_RL DNp01 VFB_jrcv07ps     <NA> <NA>
#> 2 GF, Giant Fiber        DNlt002 LTct_RL DNp01 VFB_jrcv07pu     <NA> <NA>
#>   cluster subcluster receptorType status group somaSide birthtime hemilineage
#> 1      NA         NA         <NA> Traced 10000     <NA>      <NA>        <NA>
#> 2      NA         NA         <NA> Traced 10000     <NA>      <NA>        <NA>
#>   serial serialMotif somaNeuromere somaLocation description avgLocation fruDsx
#> 1     NA        <NA>          <NA>         <NA> Giant fiber        <NA>   <NA>
#> 2     NA        <NA>          <NA>         <NA> Giant fiber        <NA>   <NA>
#>   exitNerve subclassabbr transmission tosomaLocation longTract source  soma
#> 1      None         <NA>   electrical           <NA>      none   <NA> FALSE
#> 2      None         <NA>   electrical           <NA>      none   <NA> FALSE

# use of a full CYPHER query. NB Each field relating to the neuron must be
# be preceded by "n."
bignogroup <-
  manc_neuprint_meta("where:NOT exists(n.group) AND n.synweight>5000 AND n.class CONTAINS 'neuron'")
head(bignogroup)
#>   bodyid post pre downstream upstream synweight totalNtPredictions
#> 1  28027  437 456       4868      437      5305                456
#> 2  19177  411 569       4629      411      5040                569
#> 3  20240  594 601       4447      594      5041                601
#> 4  13967  912 644       4252      912      5164                644
#> 5  17446  724 655       4455      724      5179                655
#> 6 171382  746 541       4320      746      5066                541
#>     predictedNt celltypeTotalNtPredictions celltypePredictedNt     voxels
#> 1 acetylcholine                      45044           glutamate  225825889
#> 2 acetylcholine                      18960       acetylcholine  589533979
#> 3 acetylcholine                       9470       acetylcholine  566489384
#> 4 acetylcholine                      11136       acetylcholine 1243780740
#> 5 acetylcholine                      18960       acetylcholine  750977826
#> 6 acetylcholine                      16834       acetylcholine  465918537
#>            class entryNerve            name ntAcetylcholineProb ntGabaProb
#> 1 sensory neuron   MesoLN_R SNppxx_MesoLN_R           0.5811309 0.03783341
#> 2 sensory neuron     AbNT_R   SNxx23_AbNT_R           0.9122434 0.03305553
#> 3 sensory neuron   MetaLN_R SNxx29_MetaLN_R           0.7210513 0.22840150
#> 4 sensory neuron     AbNT_L   SNxx10_AbNT_L           0.8687301 0.05167811
#> 5 sensory neuron     AbNT_R   SNxx23_AbNT_R           0.8022338 0.07676595
#> 6 sensory neuron     AbNT_R   SNxx07_AbNT_R           0.8937346 0.03987283
#>   ntGlutamateProb ntUnknownProb           origin          location locationType
#> 1      0.36905804   0.011977710 mesothoracic leg 25981,36162,29442         auto
#> 2      0.04326012   0.011440937          abdomen  25729,12900,7969         auto
#> 3      0.03870988   0.011837316 metathoracic leg 23896,22133,19804         auto
#> 4      0.06977998   0.009811865          abdomen  22095,14384,8847         auto
#> 5      0.09062613   0.030374181          abdomen  25555,14853,7796         auto
#> 6      0.05699900   0.009393622          abdomen  27032,12900,7461         user
#>   predictedNtProb prefix      rootLocation rootSide           statusLabel
#> 1       0.5811309     SN 27866,36885,22115      RHS Prelim Roughly traced
#> 2       0.9122434     SN     26421,8062,16      RHS Prelim Roughly traced
#> 3       0.7210513     SN  32867,17818,7495      RHS Prelim Roughly traced
#> 4       0.8687301     SN     25947,7375,16      LHS Prelim Roughly traced
#> 5       0.8022338     SN     26876,7774,15      RHS Prelim Roughly traced
#> 6       0.8937346   <NA>     26665,8341,29      RHS Prelim Roughly traced
#>   subclass             synonyms systematicType        target   type
#> 1     <NA>                 <NA>         SNppxx     LegNpT2_R SNppxx
#> 2     <NA>                 <NA>         SNxx23           ANm SNxx23
#> 3     <NA> ppk heat nociceptive         SNxx29 LegNpT3_R.ANm SNxx29
#> 4     <NA>                 <NA>         SNxx10           ANm SNxx10
#> 5     <NA>                 <NA>         SNxx23           ANm SNxx23
#> 6     <NA>                 <NA>         SNxx07           ANm SNxx07
#>          vfbId       modality      tag cluster subcluster receptorType status
#> 1 VFB_jrcv0lmj proprioceptive 15.08.h3      NA         NA         <NA> Traced
#> 2 VFB_jrcv0esp        unknown     <NA>      38         75         <NA> Traced
#> 3 VFB_jrcv0fm8        unknown     <NA>      43         33         <NA> Traced
#> 4 VFB_jrcv0arz        unknown     <NA>      25         68         <NA> Traced
#> 5 VFB_jrcv0dgm        unknown     <NA>      38         75         <NA> Traced
#> 6 VFB_jrcv3o8m        unknown     <NA>      22         67         <NA> Traced
#>   group somaSide birthtime hemilineage serial serialMotif somaNeuromere
#> 1    NA     <NA>      <NA>        <NA>     NA        <NA>          <NA>
#> 2    NA     <NA>      <NA>        <NA>     NA        <NA>          <NA>
#> 3    NA     <NA>      <NA>        <NA>  20453        <NA>          <NA>
#> 4    NA     <NA>      <NA>        <NA>     NA        <NA>          <NA>
#> 5    NA     <NA>      <NA>        <NA>     NA        <NA>          <NA>
#> 6    NA     <NA>      <NA>        <NA>     NA        <NA>          <NA>
#>   somaLocation description avgLocation fruDsx exitNerve subclassabbr
#> 1         <NA>        <NA>        <NA>   <NA>      <NA>         <NA>
#> 2         <NA>        <NA>        <NA>   <NA>      <NA>         <NA>
#> 3         <NA>     unusual        <NA>   <NA>      <NA>         <NA>
#> 4         <NA>        <NA>        <NA>   <NA>      <NA>         <NA>
#> 5         <NA>        <NA>        <NA>   <NA>      <NA>         <NA>
#> 6         <NA>        <NA>        <NA>   <NA>      <NA>         <NA>
#>   transmission tosomaLocation longTract source  soma
#> 1         <NA>           <NA>      <NA>   <NA> FALSE
#> 2         <NA>           <NA>      <NA>   <NA> FALSE
#> 3         <NA>           <NA>      <NA>   <NA> FALSE
#> 4         <NA>           <NA>      <NA>   <NA> FALSE
#> 5         <NA>           <NA>      <NA>   <NA> FALSE
#> 6         <NA>           <NA>      <NA>   <NA> FALSE
# }
if (FALSE) { # \dontrun{
# fetch all neurons
allneurons <- manc_neuprint_meta('where:exists(n.bodyId)')
# in theory you could do this, but it often seems to time out:
allsegs=neuprintr::neuprint_ids('where:exists(n.bodyId)', all_segments=TRUE)
} # }