Fetch neuprint metadata for MANC neurons

Fetch neuprint metadata for MANC neurons

Usage

manc_neuprint_meta(
  ids = NULL,
  conn = manc_neuprint(),
  roiInfo = FALSE,
  fields.regex.exclude = NULL,
  fields.regex.include = NULL,
  ...
)

Arguments

ids

A set of body ids in any form understandable to manc_ids

conn

Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

roiInfo

whether to include the roiInfo field detailing synapse numbers in different locations. This is omitted by default as it is returned as a character vector of unprocessed JSON.

fields.regex.exclude, fields.regex.include

Optional regular expressions to define fields to include or exclude from the returned metadata.

...

Additional arguments passed to neuprint_get_meta

Value

A data.frame with one row for each (unique) id and NAs for all columns except bodyid when neuprint holds no metadata.

See also

Examples

# \donttest{
manc_neuprint_meta("DNp01")
#>   bodyid post  pre downstream upstream synweight ntGabaProb ntAcetylcholineProb
#> 1  10000 2059 1033       4318     2059      6377  0.1707151           0.5142119
#> 2  10002 1973  934       3826     1973      5799  0.1434814           0.4596285
#>   ntGlutamateProb ntUnknownProb   predictedNt      voxels      rootLocation
#> 1       0.2175471    0.09752578 acetylcholine 38743961712 24481,36044,67070
#> 2       0.2562304    0.14065966 acetylcholine 39414880927 23217,35252,67070
#>               class entryNerve          name origin subcluster
#> 1 descending neuron        CvC DNlt002_CvC_R  brain         NA
#> 2 descending neuron        CvC DNlt002_CvC_L  brain         NA
#>            location locationType predictedNtProb prefix rootSide    statusLabel
#> 1 24481,36044,67070         neck       0.5142119     DN      RHS Roughly traced
#> 2 23217,35252,67070         neck       0.4596285     DN      LHS Roughly traced
#>   subclass        synonyms systematicType  target  type        vfbId modality
#> 1       lt GF, Giant Fiber        DNlt002 LTct_RL DNp01 VFB_jrcv07ps     <NA>
#> 2       lt GF, Giant Fiber        DNlt002 LTct_RL DNp01 VFB_jrcv07pu     <NA>
#>    tag cluster receptorType status group somaSide birthtime hemilineage serial
#> 1 <NA>      NA         <NA> Traced 10000     <NA>      <NA>        <NA>     NA
#> 2 <NA>      NA         <NA> Traced 10000     <NA>      <NA>        <NA>     NA
#>   serialMotif somaNeuromere somaLocation description avgLocation exitNerve
#> 1        <NA>          <NA>         <NA> Giant fiber        <NA>      None
#> 2        <NA>          <NA>         <NA> Giant fiber        <NA>      None
#>   subclassabbr transmission tosomaLocation longTract source  soma
#> 1         <NA>   electrical           <NA>      none   <NA> FALSE
#> 2         <NA>   electrical           <NA>      none   <NA> FALSE

# use of a full CYPHER query. NB Each field relating to the neuron must be
# be preceded by "n."
bignogroup <-
  manc_neuprint_meta("where:NOT exists(n.group) AND n.synweight>5000 AND n.class CONTAINS 'neuron'")
head(bignogroup)
#>   bodyid post pre downstream upstream synweight ntGabaProb ntAcetylcholineProb
#> 1  28027  475 456       4868      475      5343 0.03783341           0.5811309
#> 2  19177  513 569       4629      513      5142 0.03305553           0.9122434
#> 3  13967 1096 644       4252     1096      5348 0.05167811           0.8687301
#> 4  17446  868 655       4455      868      5323 0.07676595           0.8022338
#> 5 171382  894 541       4320      894      5214 0.03987283           0.8937346
#> 6  33176  377 675       5938      377      6315 0.05780571           0.4767193
#>   ntGlutamateProb ntUnknownProb   predictedNt     voxels      rootLocation
#> 1      0.36905804   0.011977710 acetylcholine  225825889 27866,36885,22115
#> 2      0.04326012   0.011440937 acetylcholine  589533979     26421,8062,16
#> 3      0.06977999   0.009811865 acetylcholine 1243780740     25947,7375,16
#> 4      0.09062613   0.030374180 acetylcholine  750977826     26876,7774,15
#> 5      0.05699900   0.009393622 acetylcholine  465918537     26665,8341,29
#> 6      0.45052716   0.014947851 acetylcholine  534616731                  
#>            class entryNerve            name           origin subcluster
#> 1 sensory neuron   MesoLN_R SNppxx_MesoLN_R mesothoracic leg         NA
#> 2 sensory neuron     AbNT_R   SNxx23_AbNT_R          abdomen         75
#> 3 sensory neuron     AbNT_L   SNxx10_AbNT_L          abdomen         68
#> 4 sensory neuron     AbNT_R   SNxx23_AbNT_R          abdomen         75
#> 5 sensory neuron     AbNT_R   SNxx07_AbNT_R          abdomen         67
#> 6 sensory neuron    ProAN_L  SNppxx_ProAN_L  prothoracic leg         NA
#>            location locationType predictedNtProb prefix rootSide
#> 1 25981,36162,29442         auto       0.5811309     SN      RHS
#> 2  25729,12900,7969         auto       0.9122434     SN      RHS
#> 3  22095,14384,8847         auto       0.8687301     SN      LHS
#> 4  25555,14853,7796         auto       0.8022338     SN      RHS
#> 5  27032,12900,7461         user       0.8937346   <NA>      RHS
#> 6 17539,32687,56294         auto       0.4767193     SN      LHS
#>             statusLabel   subclass   synonyms systematicType    target   type
#> 1 Prelim Roughly traced       <NA>       <NA>         SNppxx LegNpT2_R SNppxx
#> 2 Prelim Roughly traced       <NA>       <NA>         SNxx23       ANm SNxx23
#> 3 Prelim Roughly traced       <NA>       <NA>         SNxx10       ANm SNxx10
#> 4 Prelim Roughly traced       <NA>       <NA>         SNxx23       ANm SNxx23
#> 5 Prelim Roughly traced       <NA>       <NA>         SNxx07       ANm SNxx07
#> 6 Prelim Roughly traced hair plate hair plate         SNppxx LegNpT1_L SNppxx
#>          vfbId       modality      tag cluster receptorType status group
#> 1 VFB_jrcv0lmj proprioceptive 15.08.h3      NA         <NA> Traced    NA
#> 2 VFB_jrcv0esp        unknown     <NA>      38         <NA> Traced    NA
#> 3 VFB_jrcv0arz        unknown     <NA>      25         <NA> Traced    NA
#> 4 VFB_jrcv0dgm        unknown     <NA>      38         <NA> Traced    NA
#> 5 VFB_jrcv3o8m        unknown     <NA>      22         <NA> Traced    NA
#> 6 VFB_jrcv0plk proprioceptive 21.20.h3      NA         <NA> Traced    NA
#>   somaSide birthtime hemilineage serial serialMotif somaNeuromere somaLocation
#> 1     <NA>      <NA>        <NA>     NA        <NA>          <NA>         <NA>
#> 2     <NA>      <NA>        <NA>     NA        <NA>          <NA>         <NA>
#> 3     <NA>      <NA>        <NA>     NA        <NA>          <NA>         <NA>
#> 4     <NA>      <NA>        <NA>     NA        <NA>          <NA>         <NA>
#> 5     <NA>      <NA>        <NA>     NA        <NA>          <NA>         <NA>
#> 6     <NA>      <NA>        <NA>     NA        <NA>          <NA>         <NA>
#>   description avgLocation exitNerve subclassabbr transmission tosomaLocation
#> 1        <NA>        <NA>      <NA>         <NA>         <NA>           <NA>
#> 2        <NA>        <NA>      <NA>         <NA>         <NA>           <NA>
#> 3        <NA>        <NA>      <NA>         <NA>         <NA>           <NA>
#> 4        <NA>        <NA>      <NA>         <NA>         <NA>           <NA>
#> 5        <NA>        <NA>      <NA>         <NA>         <NA>           <NA>
#> 6        <NA>        <NA>      <NA>         <NA>         <NA>           <NA>
#>   longTract source  soma
#> 1      <NA>   <NA> FALSE
#> 2      <NA>   <NA> FALSE
#> 3      <NA>   <NA> FALSE
#> 4      <NA>   <NA> FALSE
#> 5      <NA>   <NA> FALSE
#> 6      <NA>   <NA> FALSE
# }