Fetch neuprint metadata for MANC neurons

Fetch neuprint metadata for MANC neurons

manc_neuprint_meta(ids = NULL, conn = manc_neuprint(), roiInfo = FALSE)

Arguments

ids

A set of body ids in any form understandable to manc_ids

conn

Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

roiInfo

whether to include the roiInfo field detailing synapse numbers in different locations. This is omitted by default as it is returned as a character vector of unprocessed JSON.

Value

A data.frame with one row for each (unique) id and NAs for all columns except bodyid when neuprint holds no metadata.

Examples

# \donttest{
manc_neuprint_meta("Giant Fiber")
#>   bodyid  pre post upstream downstream status    statusLabel      voxels
#> 1  10000 1033 2016     2016       5150 Traced Roughly traced 38743961712
#> 2  10002  934 1948     1948       4606 Traced Roughly traced 39414880927
#>        gaba acetylcholine glutamate    neither cropped    name      class
#> 1 0.1707151     0.5142119 0.2175471 0.09752578   FALSE 10000_R Descending
#> 2 0.1434814     0.4596285 0.2562304 0.14065966   FALSE 10000_L Descending
#>   subclass group rootSide entryNerve          position hemilineage somaSide
#> 1     <NA> 10000      RHS        CvC 24481,36044,67070        None     <NA>
#> 2     <NA> 10000      LHS        CvC 23217,35252,67070        None         
#>   serial exitNerve somaNeuromere somaLocation somaRadius description birthtime
#> 1     NA      None          <NA>         <NA>         NA Giant fiber      <NA>
#> 2     NA      None          None         <NA>         NA Giant fiber      <NA>
#>             timeStamp        type longTract  soma        top.nt     top.p
#> 1                <NA> Giant Fiber      <NA> FALSE acetylcholine 0.5142119
#> 2 2022-05-10T10:51:23 Giant Fiber      <NA> FALSE acetylcholine 0.4596285
# }