Read MANC skeletons via neuprint

Read MANC skeletons via neuprint

manc_read_neurons(
  ids,
  connectors = FALSE,
  heal.threshold = Inf,
  conn = manc_neuprint(),
  ...
)

Arguments

ids
A set of body ids in any form understandable to manc_ids

connectors
whether or not to add synapse data to the retrieved skeletons in the format used by the rcatmaid package, for easy use with rcatmaid or catnat functions. This can be done for synapse-less skeletons using neuprint_assign_connectors

heal.threshold
distance in raw units beyond which isolated fragments will not be merged onto the main skeleton. The default of 1000 implies 8000 nm for the hemibrain dataset. Use Inf to merge all fragments.

conn
Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

...
Additional arguments passed to neuprintr::neuprint_read_neurons.

Value

A neuronlist with attached metadata

Details

manc_read_neurons fetches metadata from neuprint but does not fetch synapse locations by default as this is very time consuming.

Examples

if (FALSE) {
gfs=manc_read_neurons("Giant Fiber")
gfs[,]
}