Return a Neuroglancer scene URL for MANC dataset

The default behaviour is to generate a rich neuroglancer scene with including any passed ids using the current Clio DVID node. This means that meshes should be in sync. See this slack post from Stuart Berg for more details.

manc_scene(
  ids = NULL,
  node = "clio",
  open = FALSE,
  basescene = c("2022-04-13", "2022-01-13", "2021-05-05", "2021-05-04", "2021-04-01",
    "2021-02-01"),
  server = c("clio", "appspot", "janelia"),
  return.json = FALSE
)

Arguments

ids

A set of body ids to add to the neuroglancer scene in any form compatible with manc_ids

node

A DVID node e.g. as returned by manc_dvid_node. The (recommended) default behaviour is to use the current Clio node.

open

When TRUE opens the URL in your browser.

basescene

Which neuroglancer scene url to use as a base. You can also supply your own URL.

server

Whether to use Janelia's Clio branch, the Google server (newest version of neuroglancer) or the Janelia server (required for annotation in early 2021, but now deprecated in favour of Clio). 99 should keep the default.

return.json

Whether to return a JSON fragment defining the scene or (by default) a Neuroglancer URL.

Value

A character vector containing a single Neuroglancer URL or a JSON fragment.

Details

Neuroglancer scenes can be pasted into a variety of different variants. Use the return.json to get a JSON fragment that can be pasted into any neuroglancer instance using the closed curly bracket symbol.

See slack for why https://clio-ng.janelia.org/ is the recommended base Url (chosen when server='clio').

scenes

The following scenes (named by the approximate date that we started using them) are available.

2022-04-13 Fixed synapses display and segmentation layer v.0.68.

2021-05-05 Like 2021-05-04 but with a voxelwise ROI segmentation layer copied over from 2021-04-01.

2021-05-04 Added nerves and full VNC (cell body rind) surface mesh. See Slack message from Stuart Berg. GSXEJ added the ROIs to

2021-04-01 With VNC ROIs. By April 2021 we were using Clio for annotations.

2021-02-01 In early 2021 we were using a Janelia server hosting neuroglancer that allowed annotation through a hybrid DVID backend.

The early 2021 server required authentication at https://neuprint.janelia.org/ in order to use the annotation features. I recommend logging in and out of neuprint if you still get authentication errors from Neuroglancer when attempting to use the annotation layer.

Examples

if (FALSE) {
browseURL(manc_scene())
# copy scene information with a sample neuron to the clipboard
clipr::write_clip(manc_scene(ids=13749))

# JSON fragment that could be copied into Clio
clipr::write_clip(manc_scene(return.json = TRUE))

}