Soma locations for all intrinsic neurons of the ventral nerve cord

These were generated as neuroglancer annotations by Chris Ordish of the FlyEM team and passed to Greg Jefferis on 3rd March 2021 by Pat Rivlin. These have been superseded by more dynamic data available from DVID and Clio. Therefore until the situation stabilises, it is recommended to use the manc_somapos function.

Usage

mancsomapos

Format

An object of class data.frame with 15613 rows and 8 columns.

Details

Note that these positions are raw (8nm voxel) locations.

See also

Examples

head(mancsomapos)
#>   n       X       Y       Z   kind                from                  to
#> 1 0 10001.0 36763.5 46320.5 sphere 10003, 37134, 46349  9999, 36393, 46292
#> 2 1 10009.5 29856.5 47550.0 sphere 10005, 30224, 47596 10014, 29489, 47504
#> 3 2 10024.5 34098.5 46376.5 sphere 10023, 33709, 46408 10026, 34488, 46345
#> 4 3 10033.0 30955.5 46816.5 sphere 10030, 31279, 46820 10036, 30632, 46813
#> 5 4 10035.0 31212.5 45065.5 sphere 10044, 31800, 45134 10026, 30625, 44997
#> 6 5 10038.5 33255.0 46336.0 sphere 10044, 33633, 46370 10033, 32877, 46302
#>   comment
#> 1        
#> 2        
#> 3        
#> 4        
#> 5        
#> 6        

library(nat)
points3d(xyzmatrix(mancsomapos))
plot3d(boundingbox(xyzmatrix(mancsomapos)))

3D plot

# nm
summary(xyzmatrix(mancsomapos)*8)
#>        X                Y                Z         
#>  Min.   : 62972   Min.   : 38648   Min.   :  9200  
#>  1st Qu.:146692   1st Qu.:146800   1st Qu.:112064  
#>  Median :180020   Median :227600   Median :250064  
#>  Mean   :184656   Mean   :221344   Mean   :246493  
#>  3rd Qu.:227312   3rd Qu.:301964   3rd Qu.:354780  
#>  Max.   :303640   Max.   :372024   Max.   :530840  
# microns
summary(xyzmatrix(mancsomapos)/125)
#>        X                Y                Z        
#>  Min.   : 62.97   Min.   : 38.65   Min.   :  9.2  
#>  1st Qu.:146.69   1st Qu.:146.80   1st Qu.:112.1  
#>  Median :180.02   Median :227.60   Median :250.1  
#>  Mean   :184.66   Mean   :221.34   Mean   :246.5  
#>  3rd Qu.:227.31   3rd Qu.:301.96   3rd Qu.:354.8  
#>  Max.   :303.64   Max.   :372.02   Max.   :530.8  

if (FALSE) { # \dontrun{
# find bodyid for all somata
mancsomapos$bodyid=manc_xyz2bodyid(mancsomapos)
} # }