Read draco encoded 3D meshes from tarballs (as used by Janelia FlyEM)

the low-level read_draco_meshes reads meshes from tarballs

read_manc_meshes reads the meshes for a single MANC body id.

Usage

read_draco_meshes(x)

read_manc_meshes(ids, node = "clio", type = c("merged", "supervoxels"), ...)

Arguments

x

A URL to a remote location or path to a file on disk

ids

A set of body ids in any form understandable to manc_ids

node

the dvid node containing the dataset. See manc_dvid_node

type

Whether to return a single lower resolution mesh for each body id ('merged') or a neuronlist containing one high resolution mesh for every supervoxel in a single bodyid ('supervoxels').

...

Additional arguments passed to pbsapply. to

Value

A neuronlist object containing

rgl::mesh3d objects

a neuronlist containing one or more meshes

Details

There are currently two sets of meshes available:

merged A single mesh for the entire body, in legacy neuroglancer format. This is somewhat lower resolution.

supervoxels All supervoxel meshes for a particular body, obtained as a .tar containing a .drc draco mesh for each supervoxel. Highly compressed but has more vertices.

The download size of these two objects is similar, but due to draco mesh compression the number of vertices / faces in the supervoxels meshes is 5-10x greater.

In general we recommend using the merged format, especially as this means that you can fetch many meshes at once.

Examples

if (FALSE) {
# NB you need VPN access for this
m=read_manc_meshes(23547362285)
plot3d(m)
clear3d()
plot3d(m, type='wire')
}
# Giant Fibre neurons
# \donttest{
library(nat)
gfs=read_manc_meshes(c(10000,10002))
gfs.latest=read_manc_meshes(c(10000,10002), node="neutu")
plot3d(gfs)

n10373=read_manc_meshes(10373, type='supervoxels')
nclear3d()

# supervoxels appear in different colours
plot3d(n10373)
# }